damn added a lot of stuff tbh (or not, idk anymore)

+++ backend/src/main/java/com/nutritrack/dto/FoodResponse.java
@@ -21,10 +21,10 @@ public class FoodResponse {
 
     @Data
     public static class NutritionResponse {
-        private int calories;
-        private int protein;
-        private int carbs;
-        private int fat;
+        private float calories;
+        private float protein;
+        private float carbs;
+        private float fat;
     }
 
     @Data
